经营合同查看显示合同名称 增加合同出现服务器错误 大写数字转小写数字完成

+++ src/com/jdlink/domain/QuotationItem.java
@@ -38,7 +38,7 @@ public class QuotationItem {
     /**
      * 合约量
      */
-    private int contractAmount;
+    private float contractAmount;
     /**
      * 含税单价
      */
@@ -241,11 +241,11 @@ public class QuotationItem {
         this.wastesCode = wastesCode;
     }
 
-    public int getContractAmount() {
+    public float getContractAmount() {
         return contractAmount;
     }
 
-    public void setContractAmount(int contractAmount) {
+    public void setContractAmount(float contractAmount) {
         this.contractAmount = contractAmount;
     }
